+++ SweetDelicacies/src/core/sweetdelicacies/product/Chocolate.java
@@ -4,10 +4,10 @@ import java.time.LocalDate;
 
 public class Chocolate extends Product{
     
-    float cocoaPercentage;
+    double cocoaPercentage;
     String fillingType;
     
-    public Chocolate(String name, float weight, float unitPrice, LocalDate fabricationDate, float cocoaPercentage, String fillingType) {
+    public Chocolate(String name, double weight, double unitPrice, LocalDate fabricationDate, double cocoaPercentage, String fillingType) {
         super(name, weight, unitPrice, fabricationDate);
         this.cocoaPercentage = cocoaPercentage;
         this.fillingType = fillingType;
